% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{external_and_internal}
\alias{external_and_internal}
\title{A character vector to demonstrate internal  data usage}
\format{An object of class \code{character} of length 1.}
\usage{
external_and_internal
}
\description{
A character vector to demonstrate internal  data usage
}
\keyword{datasets}
