% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{external_only}
\alias{external_only}
\title{A character vector to demonstrate external  data usage}
\format{An object of class \code{character} of length 1.}
\usage{
external_only
}
\description{
A character vector to demonstrate external  data usage
}
\keyword{datasets}
