% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_external.R
\name{print_external_from_argument}
\alias{print_external_from_argument}
\title{Print param as is}
\usage{
print_external_from_argument(x = external_only)
}
\arguments{
\item{x}{object to return}
}
\value{
param as is
}
\description{
Print param as is
}
\examples{
print_external_from_argument(x = "user_defined")
}
