% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_internal.R
\name{print_internal_from_argument}
\alias{print_internal_from_argument}
\title{Print param as is}
\usage{
print_internal_from_argument(x = external_and_internal)
}
\arguments{
\item{x}{object to return}
}
\value{
param as is
}
\description{
Print param as is
}
\examples{
print_internal_from_argument(x = "user_defined")
}
