% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_internal.R
\name{print_internal_in_body}
\alias{print_internal_in_body}
\title{Print data as is}
\usage{
print_internal_in_body()
}
\value{
external data set
}
\description{
Print data as is
}
\examples{
print_internal_in_body()
}
